This adds the original CaDiCaL_PriPro
This is the version of CaDiCaL_PriPro as it was commited to the SAT
Competition 2021. It was obtained by downloading the solver from the
official website of the SAT Competition 2021, which is
	https://satcompetition.github.io/2021
and which offered downloading all solvers submitted to the competition.
Furthermore, this commit also adds the solver description as a pdf as
submitted alongside with the solver to the SAT Competition 2021. This
solver description was also downloaded from the official website of the
SAT Competition 2021.

CaDiCaL_PriPro is a Hack of CaDiCaL v1.4.0, but unfortunately did not
qualify for the official Hack Track of the SAT Competition 2021, because
the changes to the base solver exceeded the limit of 1000 non-space
characters. Instead, CaDiCaL_PriPro regularily participated in the Main
Track, where it achieved
	 8th place in UNSAT,
	10th place in SAT+UNSAT, and
	30th place in SAT
and won the
	Special Innovation Price
due to being part of the best performing k-portfolio with minimal k
(namely, k=2) despite not winning a price in any regular Track.

This commit illustrates the changes made in order to obtain
CaDiCaL_PriPro from CaDiCaL v1.4.0.